pyproject.toml
[project]
name = "simple-bank-django"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "psycopg[binary]>=3.3.3",
]
